+++ lib/extensions.dart
@@ -3,4 +3,7 @@
 /// Exports the extension methods exposed by this package.
 library extensions;
 
-export 'src/term_similarity_extensions.dart';
+export 'src/term_similarity.dart' show TermSimilarityExtensions;
+export 'src/token.dart' show TokenCollectionExtension;
+export 'package:porter_2_stemmer/porter_2_stemmer.dart'
+    show Porter2StemmerExtension;