+++ lib/package_exports.dart
@@ -1,6 +1,0 @@
-// Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
-
-/// Exports the `porter_2_stemmer` package.
-library package_exports;
-
-export 'package:porter_2_stemmer/porter_2_stemmer.dart';