+++ lib/src/_index.dart
@@ -2,11 +2,10 @@
 // End user is granted a non-exclusive non-transferable license the ("License")
 // to use GM Consult's proprietary software (the "Software").
 
+export 'english.dart';
 export 'term_similarity.dart';
-export 'term_similarity_extensions.dart';
-export 'text_tokenizer.dart';
+export 'text_analyzer.dart';
 export 'text_document.dart';
+export 'text_tokenizer.dart';
 export 'token.dart';
 export 'typedefs.dart';
-export 'text_analyzer.dart';
-export 'english.dart';