+++ lib/src/text_document.dart
@@ -1,7 +1,7 @@
 // BSD 3-Clause License
 // Copyright (c) 2022, GM Consult Pty Ltd
 
-import 'package:text_analysis/src/_index.dart';
+import '_index.dart';
 import 'package:collection/collection.dart';
 
 /// The [TextDocument] object model enumerates properties for analysing a text