+++ lib/type_definitions.dart
@@ -1,6 +1,6 @@
 // Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
 
-/// Exports all the typedefs used in this package.
+/// Exports all the type definitions used in the `text_indexing` library.
 library type_definitions;
 
 export 'src/typedefs.dart';